Fix compiler warnings

+++ audio_buffer.hpp
@@ -7,17 +7,17 @@
 class AudioBuffer {
    public:
     bool LoadFile(std::string file_name);
-    size_t Samplerate() const { return samplerate; }
-    size_t NumChannels() const { return num_channels; }
-    size_t NumFrames() const { return num_frames; }
+    int Samplerate() const { return samplerate; }
+    int NumChannels() const { return num_channels; }
+    int NumFrames() const { return num_frames; }
     float Length() const { return static_cast<float>(num_frames) / samplerate; }
     const float* Samples() const { return samples.data(); }
     operator bool() const { return samplerate != 0; }
 
    private:
-    size_t samplerate = 0;
-    size_t num_channels = 0;
-    size_t num_frames = 0;
+    int samplerate = 0;
+    int num_channels = 0;
+    int num_frames = 0;
     std::vector<float> samples;
 };
